backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
numpy==2.2.2
statsmodels==0.14.4
yfinance==0.2.51
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.20
pandas-datareader==0.10.0
